requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorzero==2.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.56.0
freetype-py==2.5.1
fsspec==2025.2.0
gpiozero==2.0.1
gradio==5.20.0
gradio_client==1.7.2
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lgpio==0.2.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mir_eval==0.8.2
noisereduce==3.0.3
numpy==2.2.3
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pesq==0.0.4
pigpio==1.78
pillow==11.1.0
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
RPi.GPIO==0.7.1
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
svgpathtools==1.6.1
svgwrite==1.4.3
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
